Adds redeem code button visibility configuration
Allows configuring the visibility of the redeem code button via a configuration setting.

This change introduces a new configuration option to control whether the "Redeem Code" button is displayed in the store view.  The visibility is now determined by the `redeemCodeVisible` key in the configuration file.  This provides greater flexibility in customizing the user interface based on application requirements.